style(MinerrPass): remove unnecessary \" wrapping MINERR_URL
In the MINERR_SOURCE template, there is a placeholder that will be
replaced with a url that is passed into the tool.  Currently the placeholder
is \"MINERR_URL\".  The extra escaped quotes are not necessary and can be
confusing, given that the template is already a javascript function inside
a java string.